+++ PlanBuild/Client/BuildMode.cs
@@ -1,0 +1,9 @@
+namespace PlanBuild.Client
+{
+    internal enum BuildMode
+    {
+        Guide,
+        Assisted,
+        Automatic
+    }
+}